Performance improvement
Added null check on LogFile to greatly increase performance when LogFile
is not define in AppSettings

+++ Framework/Logger.cs
@@ -12,7 +12,8 @@ namespace EdiFabric.Framework
             try
             {
                 LogFile = ConfigurationManager.AppSettings["EdiFabric.LogFile"];
-                File.Delete(LogFile);
+                if (!String.IsNullOrWhiteSpace(LogFile))
+                    File.Delete(LogFile);
             }
             catch
             {
@@ -24,7 +25,8 @@ namespace EdiFabric.Framework
         {
             try
             {
-                File.AppendAllText(LogFile, text + Environment.NewLine);
+                if (!String.IsNullOrWhiteSpace(LogFile))
+                    File.AppendAllText(LogFile, text + Environment.NewLine);
             }
             catch
             {
